+++ src/pytask/execute.py
@@ -13,6 +13,7 @@ from pytask.database import create_or_update_state
 from pytask.exceptions import NodeNotFoundError
 from pytask.mark import Mark
 from pytask.nodes import FilePathNode
+from pytask.report import ExecutionReport
 from pytask.report import format_execute_footer
 
 
@@ -20,8 +21,10 @@ from pytask.report import format_execute_footer
 def pytask_execute(session):
     session.hook.pytask_execute_log_start(session=session)
     session.scheduler = session.hook.pytask_execute_create_scheduler(session=session)
-    session.results = session.hook.pytask_execute_build(session=session)
-    session.hook.pytask_execute_log_end(session=session, reports=session.results)
+    session.execution_reports = session.hook.pytask_execute_build(session=session)
+    session.hook.pytask_execute_log_end(
+        session=session, reports=session.execution_reports
+    )
 
 
 @hookimpl
@@ -41,12 +44,12 @@ def pytask_execute_create_scheduler(session):
 
 @hookimpl
 def pytask_execute_build(session):
-    results = []
+    reports = []
     for task in session.scheduler:
-        result = session.hook.pytask_execute_task_protocol(session=session, task=task)
-        results.append(result)
+        report = session.hook.pytask_execute_task_protocol(session=session, task=task)
+        reports.append(report)
 
-    return results
+    return reports
 
 
 @hookimpl
@@ -57,19 +60,13 @@ def pytask_execute_task_protocol(session, task):
         session.hook.pytask_execute_task(session=session, task=task)
         session.hook.pytask_execute_task_teardown(session=session, task=task)
     except Exception:
-        etype, value, tb = sys.exc_info()
-        result = {
-            "task": task,
-            "etype": etype,
-            "value": value,
-            "traceback": tb,
-        }
+        report = ExecutionReport.from_task_and_exception(task, sys.exc_info())
     else:
-        result = {"task": task, "value": None}
-    session.hook.pytask_execute_task_process_result(session=session, result=result)
-    session.hook.pytask_execute_task_log_end(session=session, task=task, result=result)
+        report = ExecutionReport.from_task(task)
+    session.hook.pytask_execute_task_process_report(session=session, report=report)
+    session.hook.pytask_execute_task_log_end(session=session, task=task, report=report)
 
-    return result
+    return report
 
 
 @hookimpl(trylast=True)
@@ -114,13 +111,11 @@ def pytask_execute_task_teardown(session, task):
 
 
 @hookimpl(trylast=True)
-def pytask_execute_task_process_result(session, result):
-    task = result["task"]
-    if result["value"] is None:
-        result["success"] = True
+def pytask_execute_task_process_report(session, report):
+    task = report.task
+    if report.success:
         _update_states_in_database(session.dag, task.name)
     else:
-        result["success"] = False
         for descending_task_name in task_and_descending_tasks(task.name, session.dag):
             descending_task = session.dag.nodes[descending_task_name]["task"]
             descending_task.markers.append(
@@ -135,8 +130,8 @@ def pytask_execute_task_process_result(session, result):
 
 
 @hookimpl(trylast=True)
-def pytask_execute_task_log_end(result):
-    if result["success"]:
+def pytask_execute_task_log_end(report):
+    if report.success:
         click.secho(".", fg="green", nl=False)
     else:
         click.secho("F", fg="red", nl=False)
@@ -149,19 +144,15 @@ def pytask_execute_log_end(session, reports):
     session.execution_end = time.time()
     click.echo("")
 
-    n_successful = sum(result["success"] for result in reports)
+    n_successful = sum(report.success for report in reports)
     n_failed = len(reports) - n_successful
     tm_width = session.config["terminal_width"]
 
     for report in reports:
-        if not report["success"]:
-            click.echo(
-                f"{{:=^{tm_width}}}".format(f" Task {report['task'].name} failed ")
-            )
+        if not report.success:
+            click.echo(f"{{:=^{tm_width}}}".format(f" Task {report.task.name} failed "))
             click.echo("")
-            traceback.print_exception(
-                report["etype"], report["value"], report["traceback"]
-            )
+            traceback.print_exception(*report.exc_info)
             click.echo("")
             click.echo("=" * tm_width)